add notebook to misc

+++ RTAscience/misc/deljobs.py
@@ -1,7 +1,7 @@
 import os
 import sys
 
-start = float(sys.argv[1])
+start = int(sys.argv[1])
 
 runs = 20
 for i in range(runs):
